requirements.txt
flask>=3.0.0
requests>=2.31.0
gitpython>=3.1.40
